Refactor: add phone custom hook to reuse logic #28
